Working on DCS rename prompting. fix ProjectDetailsActions unit tests
